Create a shared function for this feature
